Mostly Finalized R-Script
Here is an updated R-script and screenshots of the resulting output. The python script combines the IC50 data from GDSC and the cancer mutation hotspot data from DepMap and breaks the IC50 values into 3 classes based on value. The lowest 3rd of IC50 values are marked as "1" or most effective, the second 3rd is marked as "2" or intermediately effective, and high values are marked with "3" or least effective. The hotspot data is in binary format where a "0" represents and unmutated feature and "1" represents a mutation. These values create a kind of fingerprint comprised of over 2,400 features which identifies a specific tumor cell line.