Add files via upload
Random forest is an excellent technique that uses an ensemble learning method for both classification and regression with trees in the random forest running in parallel (Implementing Random Forest Regression in Python: An Introduction | Built In, n.d.) This means there is no interaction between the trees while building the trees. Ensemble methods combine several models to produce one with higher accuracy (Ensemble Methods in Machine Learning: What Are They and Why Use Them? | by Evan Lutins | Towards Data Science, n.d.). It is for this reason that the random forest classifier is chosen. We could apply multiple datasets to the random forest program,
which predicted much greater than the artificial neural network in python.  WARNING: UNLESS YOU HAVE A LOT OF RAM ON YOUR COMPUTER PLEASE USE A SMALLER DATASET WHEN TESTING BECAUSE IT WILL CRASH DURING THE JOBLIB PROCESS.